// _dev/src/store/modules/accounts/getters-types.ts
enum GetterTypes {
    /* PrestaShop Account */
    GET_PS_ACCOUNTS_IS_ONBOARDED = 'GET_PS_ACCOUNTS_IS_ONBOARDED',
    GET_PS_ACCOUNTS_CONTEXT = 'GET_PS_ACCOUNTS_CONTEXT',
    GET_PS_ACCOUNTS_CONTEXT_SHOPS = 'GET_PS_ACCOUNTS_CONTEXT_SHOPS',
    /* Google Account */
    GET_GOOGLE_ACCOUNT_IS_ONBOARDED = 'GET_GOOGLE_ACCOUNT_IS_ONBOARDED',
    GET_GOOGLE_ACCOUNT = 'GET_GOOGLE_ACCOUNT',
    GET_GOOGLE_ACCOUNT_TOKEN = 'GET_GOOGLE_ACCOUNT_TOKEN',
    GET_GOOGLE_ACCOUNT_MCA_LIST = 'GET_GOOGLE_ACCOUNT_MCA_LIST',
    /* Merchant Center Account */
    GET_GOOGLE_MERCHANT_CENTER_ACCOUNT = 'GET_GOOGLE_MERCHANT_CENTER_ACCOUNT',
    GET_GOOGLE_MERCHANT_CENTER_ACCOUNT_IS_CONFIGURED= 'GET_GOOGLE_MERCHANT_CENTER_ACCOUNT_IS_CONFIGURED',
    GET_GOOGLE_ACCOUNT_AUTHENTICATION_URL = 'GET_GOOGLE_ACCOUNT_AUTHENTICATION_URL',
}

export {GetterTypes as default};